Add guide to disable Philips built-in screensaver on Android TV
This commit introduces a comprehensive guide on how to disable the Philips system screensaver/overlay that appears over HDMI sources or apps. The instructions detail the steps to change a specific permission in the settings, along with optional steps to disable the Google/Android screensaver. Troubleshooting tips are also included to assist users in case of issues.